Change return type

+++ core/include/core/ICoreService.h
@@ -98,7 +98,7 @@ public:
 
     virtual std::string lastCommandDescription() const = 0;
 
-    virtual unsigned int numRevertableCommands() const = 0;
+    virtual unsigned long long numRevertableCommands() const = 0;
 
     virtual void undoLast() = 0;
 };
